// web/src/proto/product.ts
export interface Product {
	/**
	 * @generated from protobuf field: string id = 1
	 */
	id: string;
	/**
	 * @generated from protobuf field: string name = 2
	 */
	name: string;
	/**
	 * @generated from protobuf field: double price = 3
	 */
	price: number;
	/**
	 * @generated from protobuf field: int32 quantity = 4
	 */
	quantity: number;
}
// ...
class Product$Type extends MessageType<Product> {
	constructor() {
		super("pb.Product", [
			{ no: 1, name: "id", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
			{ no: 2, name: "name", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
			{ no: 3, name: "price", kind: "scalar", T: 1 /*ScalarType.DOUBLE*/ },
			{ no: 4, name: "quantity", kind: "scalar", T: 5 /*ScalarType.INT32*/ },
		]);
	}
	create(value?: PartialMessage<Product>): Product {
		const message = globalThis.Object.create(this.messagePrototype!);
		message.id = "";
		message.name = "";
		message.price = 0;
		message.quantity = 0;
		if (value !== undefined)
			reflectionMergePartial<Product>(this, message, value);
		return message;
	}
	internalBinaryRead(
		reader: IBinaryReader,
		length: number,
		options: BinaryReadOptions,
		target?: Product,
	): Product {
		const message = target ?? this.create(),
			end = reader.pos + length;
		while (reader.pos < end) {
			const [fieldNo, wireType] = reader.tag();
			switch (fieldNo) {
				case /* string id */ 1:
					message.id = reader.string();
					break;
				case /* string name */ 2:
					message.name = reader.string();
					break;
				case /* double price */ 3:
					message.price = reader.double();
					break;
				case /* int32 quantity */ 4:
					message.quantity = reader.int32();
					break;
				default: {
					const u = options.readUnknownField;
					if (u === "throw")
						throw new globalThis.Error(
							`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`,
						);
					const d = reader.skip(wireType);
					if (u !== false)
						(u === true ? UnknownFieldHandler.onRead : u)(
							this.typeName,
							message,
							fieldNo,
							wireType,
							d,
						);
				}
			}
		}
		return message;
	}
	internalBinaryWrite(
		message: Product,
		writer: IBinaryWriter,
		options: BinaryWriteOptions,
	): IBinaryWriter {
		/* string id = 1; */
		if (message.id !== "")
			writer.tag(1, WireType.LengthDelimited).string(message.id);
		/* string name = 2; */
		if (message.name !== "")
			writer.tag(2, WireType.LengthDelimited).string(message.name);
		/* double price = 3; */
		if (message.price !== 0)
			writer.tag(3, WireType.Bit64).double(message.price);
		/* int32 quantity = 4; */
		if (message.quantity !== 0)
			writer.tag(4, WireType.Varint).int32(message.quantity);
		const u = options.writeUnknownFields;
		if (u !== false)
			(u == true ? UnknownFieldHandler.onWrite : u)(
				this.typeName,
				message,
				writer,
			);
		return writer;
	}
}
/**
 * @generated MessageType for protobuf message pb.Product
 */
export const Product = new Product$Type();